[RE-938]: allow getting users from the Jira Cloud API
Cloud API is different in the places where we need to get users objects:
there are no key and name attributes for user, and you can't use
username to search for users.

I was not sure if we are ready to make our client incompatible with the
old versions, so I had to introduce a way to differntiate cloud vs
server Jira in the client sections. I use this information in the
Section\User::search method - to know which parameter to use when search
for the users.

I also changed the way we construct the User object. As there is no
"name" field anymore in the Cloud API, we can't use it in the
constructor (or otherwise you won't be able to create new objects). I
replaced it with an empty string for the backward compatibility, but we
will need to change the whole way we construct this kind of objects in
the future.

NOTE:
This changeset doesn't cover usages of the User class, it only allows to
retrieve them from the Cloud API. So some methods, like User::assign,
won't work as is.